#88 [Refactor] recordViewMany화면 체크 풀리는 문제 작업 진행중

+++ app/src/main/java/com/thequietz/travelog/data/Repository.kt
@@ -166,7 +166,9 @@ class RecordRepository @Inject constructor(
     fun deleteRecordImage(id: Int) {
         coroutineScope.launch {
             val data = recordImageDao.loadRecordImageById(id)
-            recordImageDao.delete(data)
+            if (data != null) {
+                recordImageDao.delete(data)
+            }
         }
     }
